+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Migrate the app to Yew & refactor it
 - Display inference results as log-odds or raw beliefs, at the user's choice
 - Language selection & french translation
+- Mutual information tab
 
 ## Version 0.1 -- 2019-08-02